requirements.txt
torch
transformers
SpeechRecognition
pyttsx3
openai-whisper
pydub
sounddevice
numpy
wavio
redis